+++ Classes/ImageService/ImagorImageService.php
@@ -74,6 +74,8 @@ class ImagorImageService extends ImageServiceAbstract
 			'image/gif',
 			'image/bmp',
 			'image/tiff',
+			'video/youtube',
+			'video/vimeo',
 		];
 	}