+++ Classes/ImageService/ImgixImageService.php
@@ -64,6 +64,8 @@ class ImgixImageService extends ImageServiceAbstract
 			'image/eps',
 			'image/tiff',
 			'application/pdf',
+			'video/youtube',
+			'video/vimeo',
 		];
 	}