+++ Classes/UriBuilder/ThumborFileSource.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace SomehowDigital\Typo3\MediaProcessing\UriBuilder;
 
+use SomehowDigital\Typo3\MediaProcessing\Utility\OnlineMediaUtility;
 use TYPO3\CMS\Core\Resource\FileInterface;
 
 class ThumborFileSource implements UriSourceInterface
@@ -12,13 +13,15 @@ class ThumborFileSource implements UriSourceInterface
 
 	public function getSource(FileInterface $file): string
 	{
-		return $this->build($file);
+		$url = OnlineMediaUtility::getPreviewImage($file) ?? $file->getIdentifier();
+
+		return $this->build($url);
 	}
 
-	private function build(FileInterface $source): string
+	private function build(string $url): string
 	{
-		$path = parse_url($source->getPublicUrl(), PHP_URL_PATH);
-		$query = parse_url($source->getPublicUrl(), PHP_URL_QUERY) ?? '';
+		$path = parse_url($url, PHP_URL_PATH);
+		$query = parse_url($url, PHP_URL_QUERY) ?? '';
 
 		return implode('?', array_filter([trim($path, '/'), trim($query)]));
 	}